Several fixes to correctly handle "long" attributes.

+++ src/net/sf/webcat/grader/_GradingCriteria.java
@@ -151,7 +151,7 @@ public abstract class _GradingCriteria
             (Number)storedValueForKey( "deadTimeDelta" );
         return ( result == null )
             ? 0L
-            : result.intValue();
+            : result.longValue();
     }
